Refactor: Executing targeted structural cleanup
* Evicted UI/Presentation Logic from core common to core presentation.
* Resolved External Protocol leak by moving mcp_contracts to ports directory.
* Eradicated split-brain CLI duplicates in toolkit by deleting cli.py and logger.py.
* Updated related imports and references across workspace.
* Corrected canonicalize_domain import to inline helper matching without creating domain logic.
