+++ src/coreason_manifest/core/presentation/highlighting.py
