+++ src/coreason_manifest/core/presentation/search_layout.py
@@ -13,10 +13,10 @@ from enum import StrEnum
 from pydantic import Field, model_validator
 
 from coreason_manifest.core.common.base import CoreasonModel
-from coreason_manifest.core.common.highlighting import HighlightConfig
-from coreason_manifest.core.common.templating import ParameterizedDataRef
-from coreason_manifest.core.common.transform import DataTransformSchema
-from coreason_manifest.core.common.typeahead import TypeaheadConfig
+from coreason_manifest.core.presentation.highlighting import HighlightConfig
+from coreason_manifest.core.presentation.templating import ParameterizedDataRef
+from coreason_manifest.core.presentation.transform import DataTransformSchema
+from coreason_manifest.core.presentation.typeahead import TypeaheadConfig
 
 
 class PartitionType(StrEnum):
