+++ src/coreason_manifest/core/presentation/templating.py
